requirements.txt
#prettytable==0.6.1
#python-dateutil==1.5
Django==1.5.1
django-debug-toolbar==0.9.4
